[gcc] Replace UD2 trap with JMP to next code block
GCC places a UD2 trap instruction after naked functions as a safety
measure for functions without properly returning custom epilog code.

NSEEL doesn't want to return from the functions but place many of them
in a row, so that they are executed one after the other. For that
reason find the UD2 trap and replace it with a near jump to just after
the current code block.

Since a relative JMP instruction to a target less than 127 bytes away
fits in the two bytes of the UD2 instruction, it's easily replaceable.

Assume that nothing looks like a UD2 after the actual UD2 instruction.
It's possible that some immediate value could contain the bytes of a
UD2. This code doesn't disassemble the block and check!

The alternative would have been to simply delete the UD2 and all that
comes after it, but many more complex functions have sections of the
code after the UD2 (which it properly jumps to and from). So what
comes after UD2 could be needed and cannot be thrown away.