[build][gcc] Comment out non-portable DLL copy
The DLLs are copied from the toolchain path, but that
doesn't seem to be very portable, at least I couldn't find a
way in CMake to get the current toolchain prefix, much less
the location of any dependent DLL files.

Now the build works without the toolchain file, as long as
you use the toolchain CMake to create the build.