[effect] Limit TexerII wrap mode for compatibility
The original TexerII wrapped only once (when x or y crosses +/-1) but
not again (when crossing +/-3 etc). I'd consider this a bug.

For compatibility this was reintroduced, and put behind a version
config. Every _newly_ added TexerII effect will have unlimited
wrapping, but every old preset loaded will retain the original
behavior (even when saved with this version and reloaded). There is no
UI (yet) to switch an existing effect to the new or the old version.

The first field in the effect config struct (which was unused) was
repurposed to store an integer version starting with 0 for the
original TexerII. It's fairly safe to assume that all old presets will
have a 0 here, since the config is zeroed on initialization of the
effect class. The current version is 1.

The idea for versioning is to internally add a new incremented version
macro when necessary, along with a short string aiding in the decision
whether to upgrade when saving a preset. This user interaction is not
implemented yet.