[bug] Color Map: Recreate all IDs on load
This is a sweeping but simple fix for avoiding duplicate color ids.

All IDs will now always start from the base value, and while stored,
they will not be loaded.

The map IDs are completely unused, and the color IDs also don't need
to be saved -- before too and even more so now.