[port] Move help & presets to component headers
The rationale is that these texts and preset-/example collections
should be defined and _owned_ by the components, but must be visible
to UI code as well.

They are not static (since C++ doesn't allow non-integral types for
static member values in headers), and so are duplicated at runtime
for every instance of an effect. This is very inefficient and uses
potentially way more memory (think many SSCs) but makes the code
easier to read for now.

There must be a better solution at some point. For example one could
put them back into global variables in new c_*.cpp files, but this
would create _another_ set of files.