[bug][gcc] Deal with endbr32/endbr64 from GCC>=11
GCC (or MinGW-GCC?) from version 11 onward started to generate endbr32
instructions at the beginning of functions. This is part of a safety
measure in Intel CPUs called "Control Flow Enforcement Technology"
(CET), with these instructions being used to mitigate against
hijacking jumps and and calls ("JOP", "COP"). Other instructions
mitigate against "ROP", which abuses returns.

This is mostly unintrusive, except for the way megabuf() is
implemented which replaces an 0xFFFFFFFF placeholder at a fixed offset
after the start the compiled megabuf function with the megabuf array's
start address. This assumes the naked function starts right away. The
CET protection mechanism however adds an endbr32 (or, in a future
64-bit port an endbr64) instruction to the start of the function, even
if naked. This threw off the address replacement(and resulted in a
SEGFAULT when accessing megabuf() at runtime).

The fix is twofold:
* Stop the compiler from adding endbr32 with the -fcf-protection GCC
  option. The option has two values, "branch" and "return", with the
  new default "full" meaning both of them. Use only "return" now,
  which is actually meant to say "not branch" here.
* In the megabuf code, look for endbr32/endbr64 instructions when
  replacing the megabuf address and adjust the replacement location
  accordingly.

This is a sort of "defense-in-depth" fix, because one or the other
would suffice. However, for now disable generating branching
protection with endbr32 altogether, because:
* It's unclear what could be gained by COP/JOP-ing around AVS. Who
  would want to hack it and why?
* The performance implications of adding another instruction after
  every function call are not clear. Especially with the EEL functions
  that aren't really called as functions but simply placed end-to-end
  after another and run through in succession.

However, even if you'd turn -fcf-protection to "full", the code would
still work, because megabuf now has a check for the added
instructions.