[doc] Reword README to reflect feature-development
The previous README stressed compatibility with v2.81d as the main
goal. This has been mostly (but not completely!) reached, and with
the recent larger refactoring, the closeness of the code to the
original had to be sacrificed somewhat in order to pave the way for
future ports & improvements.

Add Debian/Ubuntu/other-distro build instruction draft. The details
might be wrong, but it's a start.

Remove upstream README section. Add thank-you note to WotL.

Remove verbatim license text and reference license file instead.