[clean] Resolve signed/unsigned comparisons
Turn many loop variables etc. into unsigned integers.

For places where this doesn't work (or would cascade too far into the
rest of the code) introduce a ssizeof32(x) macro, which casts the
result of sizeof(x) to int32_t. This only works if x is known to be
small enough so its size will fit in a 31bit integer.

Define ssizeof8 and ssizeof64 as well, although they are yet unused.