[msvc] Fix function copying in EEL compiler
Function pointers, in the way they were used by the EEL compiler,
aren't working in MSVC.

MSVC moves functions around at will, so it's not guaranteed that any
function foo() will be immediately followed by its foo_end() marker
function (which is used to compute the function's size) in the binary.

Take a hint from GCC and explicitly put a marker in the form of two
UD2 instructions at the end of every function. On MSVC, use that to
compute function size, instead of the *_end() function pointers.
Note that MSVC pads functions with int3 interrupts, instead of UD2,
like GCC.

Additionally, in a debug build, MSVC puts function addresses into a
jump table, which breaks direct function binary copy by pointer
lookup.

Change function lookup to check for a jump instruction at the address,
decoding it and returning the jump target as the real address.

On MSVC a function in the binary now looks like this in the worst
case:

    0003  jmp 0x00000100        ; Jump table entry for foo()
    0008  jmp 0x00000090        ; Jump table entry for foo_end()
        ...
        ...
    0090  int3                  ; Start of foo_end(), only padding
                                ; We don't use this at all on MSVC
                                ; anymore.
        ...
    0100  mov ebp, esp          ; Start of foo()
        ...                     ; foo() body
    011a  ud2                   ; foo() end marker
    011c  ud2
    011e  int3                  ; foo() padding
    011f  int3
    0120  mov ebp, esp          ; Start of bar()
        ...

Change the EEL API for NSEEL_addfunction() and NSEEL_addfunctionex()
from taking start pointer and size of a function, to function start
and end pointers instead. This prevents the API-using code from having
to deal with the jump table, and as a bonus, simplifies the code.