[eel] Add pre-compile hook to EEL & APE interface
The AVSTrans APE hooks into the EEL compile method by overwriting the
first few instructions at function location with a call to its
preprocessing function. This way _any_ code compilation gets
intercepted and the expression string translated by AVSTrans. This
method is extremely hacky, and, while impressive and fun, not portable
at all. It also uses the hook call magic to register a post-compile
cleanup routine.

Instead, make the pre-compile and post-compile hooks an explicit part
of the EEL- and APE interface and let AVSTrans register its code
translation this way.

The post-compile hook is less expressive, because it's only used to
clean up the string memory allocated in the pre-compile hook.

Bump APE interface version to 4, which adds:
  * set_compile_hooks(precompile_hook, postcompile_hook)
  * unset_compile_hooks()
Any of the parameters to set_compile_hooks() may be NULL, but if both
are NULL the call has no effect. Use unset_compile_hooks() to clear
the hooks.