[wip] Triangle: Fix div-by-zero edge cases
The y == v1.y protection against division-by-zero (because v1.y ==
v2.y) wasn't enough. When the triangle is (completely or partly)
outside of the screen, it's possible that y == midy right away, and
the upper section of the triangle gets a line even if v1.y == v2.y,
which normally would skip right into the second half. This effect
came into play when either the lower or the upper limit for the
initialization for y and midy are hit in max(0, min(h, v1.y))) etc.

To fix this issue make sure that the limits for midy are lower than
those for y. This makes sure the y <= midy abort condition of the
first for loop is false when the triangle's upper edge is horizontal
and outside of the screen.

Also move out the y == v1.y check to before the loop, so it's not
evaluated every line.