[bug] ColorMap: Fix adding new first color
When adding a new _first_ color the actual adding was skipped, because
the case where the new position is before the previous first position
wasn't considered in the insertion algorithm. This caused the last
color entry of the new map to be uninitialized which lead to crashes
when accessing that color later.

Check for the `new_position < first_position` case explicitly before
inserting the rest of the colors.