[bug] Colormap: Fix disabled first map selected
On config load the current map was not set. It is now set to the first
enabled map. If all maps are disabled, the first map is used for
rendering (and by that it is virtually enabled, although the state and
checkbox still stay disabled).

When opening the UI the first map was always selected (like in the
original). Now the current map is displayed, which was set on load or
whenever the user previously changed the current map.

(And rename a too-long variable name.)