[port] Explicit MAX_PATH & split off LEGACY_SAVE_PATH_LEN
`MAX_PATH` is defined in windows.h but the goal is to reduce use of
that header in the backend.

While `MAX_PATH` could conceivably be enlarged in the future, fix
`LEGACY_SAVE_PATH_LEN` at 260 characters, because that's what the
preset file format uses for many components that save file paths as
part of their configuration. Now the individual components have to be
changed to use this macro instead of `MAX_PATH`.

Move all of this into r_defs.h because it's not platform-dependent
anymore.