[build] Use add_compile_definitions() in CMake
According to the CMake documentation `add_definitions()` is intended
to be used for compile-time define flags but is in fact a generic
method that can add any flag.

It's superseded by `add_compile_definitions()`, so use that. It also
removes the -D in front of all flags, which is nice.