[port] Add new libavs API
Add new C API code to separate the rendering part of AVS from the UI &
Winamp plugin part. This will allow independent usage of the AVS core
rendering functionality in the future. The advantages are:

  - AVS becomes less platform-dependent, without any UI code directly
    attached. This was more or less a requirement for a feasible Linux
    port.
  - AVS becomes much more testable, with a library that can render
    preset frames as needed and in a reproducible fashion.
  - AVS can gain new and different UIs.
  - The API is not effect-specific. This means that future new effects
    don't have to define their own UI, they only need to specify their
    parameters (they have to anyway) and any UI should be capable of
    presenting a generic list of parameters.

Include a C++ wrapper around the C API.

---

The API is split in two parts: The basic player API, defined in avs.h,
and the editor API defined in avs_editor.h. This way any users of the
library that only want to render presets from audio can do this with
the much simpler basic API, without having to navigate through the
larger editor API.

The basic API allows rendering frames, optionally passing in audio,
and loading and saving presets, both in a (future) JSON format as
well as the now termed "legacy" binary format. Most of the
implementation of the basic API only consists of stubs for now.

The editor API exposes an index of available Effects from which new
Components may be created. Components are instances of their Effect.
Presets consist of a tree of Components (like they always have).
Effects (and so Components as well) have Parameters, which can be of
one of a handful of basic types.

Usage of the editor API is designed around two points of entry: The
first involves querying the available library of effects, so the UI
user can add them to the preset. To view and edit the current preset,
retrieve the current tree of Components, query a Component's Effect's
Parameters and show their metadata (group, name, description, etc.)
and values, and set new values.

---

On the interior side of the API change the way that effects are
defined. Currently, all effects are C++ classes that inherit a common
rendering interface from `C_RBASE` but are otherwise completely
independent when it comes to parameters. So far the UI code making
use of them is written with access to the class interface. With the
generic editor API this is no longer true and effects will have to
provide a Config/Info/Effect structure to support the API described
above. Introduce `Effect_Config`, `Effect_Info` and `Effect` base
classes to be inherited by new-style effects.

Many effects already have some sort of config struct and the rest can
easily be adapted. But all effects now have to specify the additional
`Effect_Info`-derived metadata struct as well:
An `Effect_Info` struct holds all the metadata needed for parameter
introspection. At its core it records, for each parameter in the
effect's config, the type of the parameter and its offset location in
the config so that it can be accessed and modified through the generic
API later. For each parameter a list of other things are recorded as
well: Name, description, selectable options, value limits for numerical
parameters, or callback functions to be called when the value is
changed through the API.

Additionally, mostly for Color Map but also some other effects,
parameters may be grouped into variable-length lists of sub-
parameters. Color Map for example will have nested parameter lists:
A list of maps, each containing a list of colors, with each color
having both position in the map and it's actual color value. List
elements can be added, removed and moved around through the API, and
effects can listen for these changes through callbacks. List lengths
can be limited in both directions.

---

Allow several concurrent AVS instances, each with their own preset
tree. But this part of the new code isn't used yet. First all effects
must be ported to the new Config/Info/Effect structure. After that the
old cfgwin & rlist structures can be ported to the new model around
instances and the rewritten `Effect` tree.