[port] Add Legacy_Effect_Proxy to help transition
Legacy_Effect_Proxy is a class that dispatches between new- and
old-style effects, while they're being refactored.

Change rlib to register new-style effects along with the old.
Change r_list to handle the proxy class on component creation.
Change cfgwin to handle the proxy class when displaying the UI for an
effect.