[bug] Fix window destroy race on preset-change
When changing presets without transition while a component is selected
and its UI is visible, the UI destroy calls are issued from the UI
thread after the render components have already been cleaned up in
the render thread. This didn't cause any issues so far because
(apparently) no window handler accessed re-allocated memory on
destroy. It surfaced now when refactoring ColorMap's UI code.

This is a quick-and-easy fix because the preset-change code will have
to be adapted to the new component-tree model soon.