[api] Allow setting values on list-element-add
On parameter list element creation, a handler may be run. Before this
change the handler would have been run on the list with only the
default-initialized new element. Setting any values for the new
element would have run after the on-add handler. This is a problem
when the handler moves elements around (such as sorting) and the
API-using code can now no longer be sure which one the new element is.

There were several options for solving this problem:
1. Edit values on creation of the new element.
2. Explicitly disable or delay running the handler somehow while the
   element values are set.
3. Make a two-step API for adding elements, with "begin" and "end"
   methods.

Both 2 and 3 would introduce state into the API call itself, and add
potential for error: What limits would the delay have in option 2?
What if the "end" method is not called in option 3?

The trade-off for option 1 is that it introduces the somewhat
redundant and (more importantly) inconsistent `AVS_Parameter_Value`
and `AVS_Value` structs, which are only used for the
`list_element_add` method and not for all the other `set` methods.
This could be a possible source of confusion for learners of the API.

It also introduces unions into the API-using code (even through the
C++ wrapper), whose initialization looks ugly, IMO. C++17 introduces
std::variant, which would make this slightly less ugly, but avs_cpp.h
is supposed to stay C++11 for now.