[api] Add int-, float- & color array getters
Reading large amounts of values from the API with the normal getters
would quickly become inefficient because the API would have to resolve
its handles for every value. Instead add read-only array getters for
the 3 value types that arrays are likely to consist of: numbers and
colors.

If other value types are needed in the future they are easily added.

_Writing_ large amounts of values through the API is another task,
which may not even be needed. At least none of the current AVS effects
would make use of it.

Add a thin array-with-size-class layer to the C++ API wrapper,
`Readonly_Array`, as the array getter methods' return type.