[feature] Forward SHOWWINDOW messages to effect UI
When the editor window is closed it is in fact just hidden. The
SHOWWINDOW event message is not forwarded to child dialogs. But the
ColorMap UI needs to know when the config window is visible to enable
and disable on-beat map changing. Previously it made use of a timer to
check periodically.

This seems wasteful, so forward the SHOWWINDOW event to the currently
selected effect dialog. This enables effects' dialog handlers to deal
with visibility when it changes. The refactored ColorMap will make use
of this.