[port][effect] New-style ColorMap
This effect will probably be the most involved change of all the
effects. ColorMap makes extensive use of various types of parameters,
including parameter lists (and was the reason for having them in the
first place), -actions and -arrays.

Some notable changes:
- Baked maps are now part of each map config, instead of being a
  separate list of maps.
- Code for loading & saving a single map is moved from the UI into
  action parameters. Only the file chooser code remains in the UI
  code.
- Remove the use of a timer to check UI visibility to handle map
  changing, and use the forwarded WM_SHOWWINDOW message instead.

If, for some reason, this is the first new-style effect change you're
looking at, then read "07caba4 [port][effect] New-style Fadeout"
first, for a similar but much simpler changeset.