[build][gcc][msvc] Pin C++ compiler std to C++14
C++14 is the maximum level of C++ features we want to use for now.
This cutoff point is rather arbitrary, but for now we don't really
need anything more fancy than this. Pinning it in the flags ensures
no code is added accidentally that requires newer features.

Most of the code is C++11 but we need multiline constexpr functions
for better parameter struct initialization code (in a future commit),
which are only available from C++14 onward.

There's a section in files.cpp which uses std::filesystem which is
only available from C++17 onward, but it's unused at the moment. And,
at least on the GCC/MinGW implementation, it performed far worse than
both Win32's FindFile and Linux' ftw and glob APIs.