[msvc] Fix min/max not defined without windows.h
It turns out, that min/max are not MSVC builtins. They are defined in
windows.h but not on MinGW, which led to the appearance that the
compiler difference was to blame. Now with lots of files not including
windows.h anymore missing min/max errors popped up on MSVC.

`NOMINMAX` is an ifdef to prevent definition of the min & max macros
in windows.h. Set `NOMINMAX` project-wide.

Rename platform_gcc.c to platform.c, since it only contains
definitions for min() and max(). Move the declaration out of the
compiler ifdefs in platform.h.