[msvc] Use .def file for DLL function exports
MSVC offers two methods for exporting functions from a library:

  1. Add __declspec(dllexport) to the function declaration, or
  2. List to-be-exported function names in a def file.

We previously tried 1. but only in the .cpp files, not in the header.
For MinGW-GCC this is sufficient, but MSVC also requires the declspec
on the function declaration in the header file. This would clutter up
the header file too much in my opinion, so we go for option 2. which
CMake makes easy enough.