[port] Add new Programmable_Effect base class
This child template of the `Configurable_Effect` template will take
the place of `ComponentCodeBase` and help with de-duplicating code for
EEL-using effects.

Instead of a member variable this is an effect base class template
that takes (when compared to `Configurable_Effect`) an additional
template parameter, a `Variables` child struct for EEL code variables.

It automatically creates and uses a lock for thread safety against
concurrent compile & exec calls.

Its `Code_Section` class is slightly more complex than the other
version in that it holds references to the source code string, the EEL
context and the lock. The EEL context is a pointer _reference_ even
though the actual pointer address does not change during the lifetime
of the instance now. But future code very well may delete and create
a new EEL context during the effect instance's lifetime for some
reason. So let's play it safe here now. In contrast I don't expect the
locking code to change in this way, hence the plain pointer there.