[clean] Fix comparison warning w/ 0 effect params
This is a rather crude fix for the warning, but it's probably the
simplest one. If an effect has zero parameters, the compiler
(correctly) complains that the comparison in the for loop condition
is always false.

However we don't want to have a) another specific macro for this case,
or b) check this case explicitly (since the for loop does exactly the
same thing). Instead we make the index variable signed (which silences
the warning) and larger (to avoid overflow, however unlikely).