[port][effect] New-style Scatter
Notably this is the first effect that has its `enabled` member removed
in favor of the one from the base class.

Rename from c_-/r_-/g_scat to e_-/g_scatter. I am fairly certain the
original authors were well aware of the double meaning and found it
hilarious. Sorry to be a party pooper.