[feature] Load FFmpeg/libAV DLLs at runtime
Before, libavformat, libavcodec & swscale were linked to AVS and it
wouldn't start without them present. This is bad because video
playback is far from essential for AVS and FFmpeg/libAV are no
lightweight dependencies.

Load the libraries and methods at runtime instead, which in case of
error allows us to keep the rest of AVS working -- without video
support of course.

Move libav code into separate `video_libav.h`/`.cpp` files, because
video.cpp was getting too large with all the function pointer
declarations, retrieval and definitions.

The libav DLLs are loaded only once for all instances of `AVS_Video`,
but never unloaded. This is of course not desirable but since they
should remain loaded for the entire lifetime of vis_avs (and a future
libavs) this is not critical. Add a TODO to create a way to register
cleanup methods when AVS itself exits.