[port][effect] New-style AVI
Use the new `AVS_Video` for playback. Support many more video formats
instead of just AVI.

Add a resampling setting to the UI. This is a departure from the
strict compatibility of the old AVS. Cannot be saved at the moment
because the save format is unchanged.

Remove video-for-windows (vfw) include and linking.