[bug] TexerII: Restore adj-blend for scale mode
Adjustable blend with resizing on acted more like Additive blend.

The Adjustable blend slider value is _still_ interpreted the wrong way
around, like in the original TexerII. This would be easy enough to
fix, but it would break presets which where created with the bug
present.

In the future the slider interpretation should be flipped, while also
loading legacy presets correctly. There are at least two ways to a
solution:

  - Fix the slider for TexerII in general. During legacy preset import
    check for the presence of "Set Render Mode" effects before any
    TexerIIs and switch the slider in these. But if, and only if, no
    other render effects (e.g. SuperScope) are affected by the mode
    change as well. In the latter case insert additional "Set Render
    Mode" instances before and/or after TexerII instances as needed.
    More complicated cases might be possible.

  - Add a TexerII version marker which applies the correct behavior
    only for new instances and loads legacy presets with the old one.
    The machinery for this is already in place for wrapping. This
    makes it very easy to add. Drawbacks: First, it only postpones the
    decision until the moment of saving a legacy preset in the new
    file format. Second, it carries the bug forward into the future
    because it's not contained to just the legacy import anymore but
    extends to a switch in the rendering code.

The first way is probably the way to go.