[bug] Save old effects' strings with null byte
AVS had a homegrown string class: `RString`. Its `length` included its
null-terminator because it was also used for memory allocation.
Effects using this class as config parameter types saved those
parameters as length-prefixed string which still included the null-
terminator in both the saved length as well as the actual string.

Strings loaded in this way included an extra null byte at the end
because `std::string::assign(char*, size_t)` does not stop at null
bytes, but does as it's told and creates a string of the given size.
Saving the string added another byte to the length (actually an
unrelated bug) which added another null byte every time the preset
was saved again.

This didn't lead to errors when loading the preset with this version
of AVS, but garbage appeared at the end of strings when opening such
a preset in the original AVS.