[api] Add global config parameters to Effect class
A few effects have config parameters that are "global" to all
instances in the presets. Namely "AVSTrans" and "MultiDelay" and to a
lesser extent "Triangle".

Previously these global config parameters where implemented using
static members and/or global variables. There are two issues with
this: First, code duplication when each effect has to implement its
own version of this. Second, a future feature of libavs will be to
support multiple AVS instances at once, i.e. rendering multiple
presets. In that case static members are "too global" and they would
affect components in other instances, which they should not.

To address this add a per-AVS-instance "global" config construct to
`Configurable_Effect`. This is a list of `Effect_Config`s (just like
the regular component config) keyed by the AVS instance handle.
Components can now use their new `global->config` member which is a
pointer to the AVS-instance-wide effect config. `global->instances`
contains a list of all components currently sharing the same config.

Add new `is_global` member to the API's `AVS_Parameter_Info` struct.

A set of changes to the API backend code is necessary to support this:

  - For each P_* function for defining parameters (except for Actions)
    add a new P_*_G version which declares a parameter as global.
  - Add a new template parameter to `Configurable_Effect`:
    `Global_Config_T` (with the empty `Effect_Config` as default).
    Effects which don't need global configuration aren't affected
    because of the default, but also have an invalid `global` member!
  - Change the get/set code in effect.h and effect_info.h to dispatch
    to the component config or the global config based on what the
    parameter declared. While on this, clean the parameter edit code
    in effect.h a bit and remove the weird `Config_Parameter` struct,
    at the cost of separately looking up the parameter handle and then
    the address (doing a merged lookup was probably premature
    optimization anyway -- most effects have very few parameters).
  - Pass the containing AVS instance handle when initializing
    components. For now (while still using `RLib` and while
    `AVS_Instance` is still not used) use 0 as placeholder.

Change avs_editor.cpp as well as the UI handlers for "ColorMap",
"RotStar", "SuperScope" and "Simple" to account for the change in
signature of the list edit functions.