[bug][clean] Better parameter-handles-for-API
Each effect's Info class has a static array of all its parameters,
including the necessary metadata, like their handles, names, types,
etc. The API is supposed to hand out a list of just the handles for
the parameters. So a second array is kept, `parameters_for_api`,
which is filled with just the handle from each parameter in the first
array. The API uses `get_parameters_for_api()` to retrieve that list
and hand it to the caller.

With GCC there was a very weird linker warning for ColorModifier and
DynamicDistanceModifier, which looked like this:

    g_colormodifier.cpp.obj: duplicate section
    `.data$_ZZNK18ColorModifier_Info22get_parameters_for_apiEvE18parameters_for_api'
    has different size

The only difference was that those two are the only effects (among the
ones ported so far) to have 7 and 6 parameters respectively, whereas
all the other effects have either less or more than that. Curiously,
increasing the declared array length (the number of actual elements
didn't matter!) to 8 makes the warning go away, which makes it sound
like some sort of memory aligment bug in the linker or compiler.

Also, IIRC, when testing whether AVS would compile with Clang it
complained about the function-static `parameters_for_api` array in
`get_parameters_for_api()` which had non-literal (although constexpr)
size.

As a result, add a new `Parameter_Handle_List` struct template as a
fancy C++14 way of expressing a fixed compile-time array which can be
initialized by a function using data from another static array instead
of literal initialization, i.e. braced-init-list, `{1, 2, 3}`.

The handles-only array is now an instance of this struct, which
contains the actual array. The filling loop moves into the struct's
constructor which is constexpr, i.e. can run at compile-time. Before
this the secondary array was actually (uselessly) re-written every
time `get_parameters_for_api()` was called.

This solves the linker warning, reduces runtime overhead, and possibly
fixes the Clang compiler issue. All at the cost of a few lines of less
intuitive code.