[clean] Exclude apesdk & evallib from clang-format
"avs/vis_avs/apesdk" is the APE authoring tutorial, which is unused
but not yet removed. "avs/vis_avs/evallib" is an older version of the
contents of "avs/ns-eel" but with flex and bison source files present,
also unused but not removed yet.

These directories won't have changes and don't need reformatting.