[bug] Fix parameter-list-add negative index wrap
`avs_parameter_list_element_add()` allows negative indices to place
the new element at some offset from the _back_ of the list. While the
value-setting code accounted for these negative parameter list
indices, as did `list_remove()` -- `list_add()` did not, and treated
all negative indices like "at the very end". Fix this by wrapping
negative indices once and, if still negative, fail.

Note that positive out-of-bounds indices never fail but place the new
element at the end of the list instead. That way one can use an
enormous index to quickly add to the end without querying list length.

It would be nice to have those fail as well, like other places of the
code but it's not worth the extra API call (to query the exact length
with `avs_parameter_list_length()`) for this, presumably, very common
case. Another option would have been to make -1 mean "end-of-list",
and -2 "one-before-last" and so on, but this would have introduced a
very confusing off-by-one to negative indices that's not happening
anywhere else.