[port][effect] New-style Colorfade
Fix a bug where the 2nd and 3rd slider have switched behavior when
easing down from their on-beat value. This made the on-load- and
runtime values for those two sliders different. Introduce a version
marker like for TexerII in the legacy file format in an unused byte
in the config.

Add extensive comments explaining how the channel-fader-switching
lookup table works.