[clean] Make find_files_by_extensions' args const
When `find_files_by_extensions()` is run in the background it copies
some of its arguments, namely `path`, `extensions` and `callback_data`
to allow the calling code to deallocate them. By making the normal
case const, this needs a little refactoring for the background case,
by wrapping the find function in another wrapper with non-const
arguments which are freed afterwards.