[api] Add new Resource parameter type
Select-type parameters use an index as value and operate under the
assumption that the list is constant over time, so that loading a
preset with a saved index will refer to the same select value that it
was saved with.

This assumption is arguably false for file resources. The list of
available files is different on every system and changes over time.
To account for this introduce a parameter type similar to
AVS_PARAM_SELECT, but whose underlying data type is not integer but
the option string value: AVS_PARAM_RESOURCE.

It requires only minor changes, except for the string comparison loops
necessary when setting a new value. We don't want to be able to set a
value that's not in the options list.