[ui][feature] Add a missing resource to dropdown
If the resource value is not found in the options list it's helpful to
still display the filename in the combobox, so one can look for it
elsewhere. To make it clear what's happening, add a "[missing!]"
disclaimer after the filename.

Fix a small warning about uint -> int implicit conversion.
Fix another dormant missing type `Parameter` error by including
`effect_info.h` in `g__defs.h`.