[ci] Merge 3 GH-CI files into one
This reduces the amount of actions in the actions list. So far there
were always 3 for every commit, but as soon as you click on it,
there's an overview with only one item that you have to click through
anyway. So let's declutter the actions list, and have all checks and
builds for a commit in one place: the run detail view.