[refactor] Add == & != operators for effect types
This allows one to quickly determine the type of an Effect- or an
Effect_Info instance by comparing to a known Effect_Info subclass:

    Effect* my_effect = ...
    if(my_effect == EffectList_Info()) {
        printf("This is an Effect List\n");
    } else {
        printf("This is a regular effect\n");
    }

or something to that effect. The "Effect == Effect_Info" comparison
looks slightly off, because they're different types and abstractions,
but the above is still much nicer to read than

    if(my_effect->get_info() == EffectList_Info) { ... }

But the latter is also valid now.