[platform] Return func-ptr, not data-ptr, on library_get()
`void*` and `(*)()` are fundamentally different in C. The first type
points to a value of an opaque/unknown type, and the second points to
a callable with (possibly) unknown signature.

Casting data pointers to function pointers is invalid in C, so use an
actual function pointer type here.

Also, clean up the code in SVP to use function pointer type and a more
readable variable name.