[refactor] Use location-aware comptime handle macro
Using #__LINE__ or #__FILE__ (i.e. the strings of these values) inside
a macro would result in those values being those of the macro
definition location instead of the usage location.

But of course for every evil macros create they hold a solution
through deeper magicks: Using a twice-nested macro will apparently
confuse the macro expansion just enough that it will expand
__LINE__ and __FILE__ not until the final location.

While we're at it, add a more generic macro COMPTIME_HANDLE_FROM(),
that takes a string argument which is appended to the location string
to be used for the FNV1a hash integer.