[refactor] Add copy & move to (Configurable)Effect
Every effect gains an individual `clone()` implementation because copy
constructors are not inherited. But on copying a derived class the
base classes' copy constructors are implicitly called. So we don't
need to repeat the copy (and move) constructors & assignment
operators -- only the clone method. We want this so the return type is
not `Effect*` but a pointer to the actual effect's class.