[bug] Swap framebuffers before copying to output
The primary framebuffer (given by external caller of
`avs_render_preset()`) may not be the last framebuffer rendered to.
But it's supposed to be the starting primary framebuffer for each
frame. If primary and secondary buffers are swapped after the preset's
effects have been all rendered, then we don't just need to copy the
contents to the external framebuffer, but also swap it back into first
position for the next frame.