[bug] Add final framebuffer copy on odd swap parity
Each effect may either render to the first or the second framebuffer.
If it renders to the secondary buffer, the buffers need to be swapped
afterwards (only in name, i.e. pointer) for the next effect-render
call.

The initially-first framebuffer is the pointer to the output
framebuffer memory provided from the outside. If after the last
effect that pointer ends up swapped to the secondary position, the
actual final image is in the other framebuffer and a final copy is
needed.

This got lost in the Root/EffectList split refactor.